fix ignores
