+++ OpenAlprWebhookProcessor.Server/Features/Ignores/Commands/DeleteIgnore/DeleteIgnoreCommand.cs
@@ -1,0 +1,15 @@
+using Mediator;
+using System;
+
+namespace OpenAlprWebhookProcessor.Features.Ignores.Commands.DeleteIgnore
+{
+    public class DeleteIgnoreCommand : ICommand
+    {
+        public Guid Id { get; set; }
+
+        public DeleteIgnoreCommand(Guid id)
+        {
+            Id = id;
+        }
+    }
+}
