+++ OpenAlprWebhookProcessor.Server/Features/Ignores/Queries/GetIgnores/GetIgnoresQuery.cs
@@ -1,0 +1,9 @@
+using Mediator;
+using System.Collections.Generic;
+
+namespace OpenAlprWebhookProcessor.Features.Ignores.Queries.GetIgnores
+{
+    public class GetIgnoresQuery : IQuery<List<IgnoreDto>>
+    {
+    }
+}
