+++ OpenAlprWebhookProcessor.Server/Features/LicensePlates/Commands/UpdatePlateNumber/UpdatePlateNumberCommand.cs
@@ -27,3 +27,4 @@ public class UpdatePlateNumberCommand : ICommand
 
 
 
+
