+++ OpenAlprWebhookProcessor.Server/Features/LicensePlates/Commands/UpdatePlateNumber/UpdatePlateNumberCommandHandler.cs
@@ -33,20 +33,3 @@ public class UpdatePlateNumberCommandHandler : ICommandHandler<UpdatePlateNumber
     }
 }
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
