--- OpenAlprWebhookProcessor.Server/Features/Settings/Commands/DeleteIgnore/DeleteIgnoreCommand.cs
@@ -1,15 +1,0 @@
-using Mediator;
-using System;
-
-namespace OpenAlprWebhookProcessor.Features.Settings.Commands.DeleteIgnore
-{
-    public class DeleteIgnoreCommand : ICommand
-    {
-        public Guid Id { get; set; }
-
-        public DeleteIgnoreCommand(Guid id)
-        {
-            Id = id;
-        }
-    }
-}
