--- OpenAlprWebhookProcessor.Server/Features/Settings/Commands/UpdateIgnore/UpdateIgnoreCommand.cs
@@ -1,15 +1,0 @@
-using Mediator;
-using OpenAlprWebhookProcessor.Features.Settings.Queries.GetIgnores;
-
-namespace OpenAlprWebhookProcessor.Features.Settings.Commands.UpdateIgnore
-{
-    public class UpdateIgnoreCommand : ICommand
-    {
-        public IgnoreDto Ignore { get; set; }
-
-        public UpdateIgnoreCommand(IgnoreDto ignore)
-        {
-            Ignore = ignore;
-        }
-    }
-}
