--- OpenAlprWebhookProcessor.Server/Features/Settings/Queries/GetIgnores/GetIgnoresQuery.cs
@@ -1,9 +1,0 @@
-using Mediator;
-using System.Collections.Generic;
-
-namespace OpenAlprWebhookProcessor.Features.Settings.Queries.GetIgnores
-{
-    public class GetIgnoresQuery : IQuery<List<IgnoreDto>>
-    {
-    }
-}
