--- OpenAlprWebhookProcessor.Server/Features/Settings/Queries/GetWebhookForwards/GetWebhookForwardsQuery.cs
@@ -1,9 +1,0 @@
-using Mediator;
-using System.Collections.Generic;
-
-namespace OpenAlprWebhookProcessor.Features.Settings.Queries.GetWebhookForwards
-{
-    public class GetWebhookForwardsQuery : IQuery<List<WebhookForwardDto>>
-    {
-    }
-}
