+++ OpenAlprWebhookProcessor.Server/Features/WebhookForwards/Queries/GetWebhookForwards/GetWebhookForwardsQuery.cs
@@ -1,0 +1,9 @@
+using Mediator;
+using System.Collections.Generic;
+
+namespace OpenAlprWebhookProcessor.Features.WebhookForwards.Queries.GetWebhookForwards
+{
+    public class GetWebhookForwardsQuery : IQuery<List<WebhookForwardDto>>
+    {
+    }
+}
