+++ OpenAlprWebhookProcessor.Server/Features/WebhookForwards/Queries/GetWebhookForwards/WebhookForwardDto.cs
@@ -1,0 +1,19 @@
+using System;
+
+namespace OpenAlprWebhookProcessor.Features.WebhookForwards.Queries.GetWebhookForwards
+{
+    public class WebhookForwardDto
+    {
+        public Guid Id { get; set; }
+
+        public Uri Destination { get; set; }
+
+        public bool IgnoreSslErrors { get; set; }
+
+        public bool ForwardGroupPreviews { get; set; }
+
+        public bool ForwardSinglePlates { get; set; }
+
+        public bool ForwardGroups { get; set; }
+    }
+}
