+++ OpenAlprWebhookProcessor.Server/Features/WebhookForwards/WebhookForward.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace OpenAlprWebhookProcessor.Features.WebhookForwards
+{
+    public class WebhookForward
+    {
+        public Guid Id { get; set; }
+
+        public Uri Destination { get; set; }
+
+        public bool IgnoreSslErrors { get; set; }
+
+        public bool ForwardGroupPreviews { get; set; }
+
+        public bool ForwardSinglePlates { get; set; }
+
+        public bool ForwardGroups { get; set; }
+    }
+}
