+++ openalprwebhookprocessor.client/src/app/settings/settings.service.ts
@@ -3,7 +3,7 @@ import { Injectable, inject } from '@angular/core';
 import type { Observable } from 'rxjs';
 import type { Alert } from './alerts/alert';
 import type { Camera } from './cameras/camera';
-import type { Ignore } from './ignores/ignore';
+
 import type { Agent } from './openalpr-agent/agent';
 import type { AgentStatus } from './openalpr-agent/agentStatus';
 import type { AgentVideoStreams } from './openalpr-agent/videoStream';
@@ -54,22 +54,6 @@ export class SettingsService {
 
   startAgentScrape(): Observable<null> {
     return this.http.post<null>('/api/settings/agent/scrape', null);
-  }
-
-  getIgnores(): Observable<Ignore[]> {
-    return this.http.get<Ignore[]>('/api/settings/ignores');
-  }
-
-  deleteIgnore(ignoreId: string) {
-    return this.http.delete(`/settings/ignores/${ignoreId}`);
-  }
-
-  upsertIgnores(ignores: Ignore[]) {
-    return this.http.post('/api/settings/ignores', ignores);
-  }
-
-  addIgnore(ignore: Ignore) {
-    return this.http.post('/api/settings/ignores/add', ignore);
   }
 
   addAlert(alert: Alert) {
